requirements.txt
asgiref==3.7.2
aspose-words==23.6.0
Babel==2.12.1
Django==3.2.19
docxcompose==1.4.0
docxtpl==0.16.7
et-xmlfile==1.1.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
numpy==1.25.0
openpyxl==3.1.2
pandas==2.0.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
XlsxWriter==3.1.2
